Test environment_variables_are_passed_one_by_one on Windows
This enables the `environment_variables_are_passed_one_by_one` test
for `gix-command` on Windows. It was marked to run on Unix-like
systems only. But it runs and passes on Windows, requiring no
changes. (Possibly it hadn't passed at the time it was introduced.)

The `command_may_be_shell_script` call is sufficient even though it
does not disable manual argument splitting, because the command
used in this test contains `$` (to expand variables in the shell)
and is thus already ineligible for manual argument splitting.