The set-display script needs to be sourced, not run
This removes the executable bit and shebang, names it `.sh` as is
traditional for a file being sourced, and sources it.

The reason it needs to be sourced is that it is using a symlink
that goes through an entry specific to the current proces under
`/proc`. This is done because the symlink is one of the symlinks in
the process filesystem that has special semantics: it refers to
stdout for the process, even when stdout is not a file that could
otherwise be accessed with a path on disk. (We need that because
the stdout and stderr streams in a GitHub Actions step go to an
unnamed pipe object.) But if it is run as a script then the new
shell instance that runs the script is the current process, which
then goes away, breaking the symlink (or causing it go to the wrong
place if another process is created that reuses the old PID).