Add some tests for finding associated executables
The tests are not limited to finding `sh`, but this may still not
be ready for use in finding other commands, for the commented
reasons. These tests are a step toward that, but they are mainly
to make sure the search works as expected both when the looked-for
command is and is not found in one of the searched `bin` dirs.

This is to say that, in the tests, the commands that can be found
but in `usr/bin` rather than the first choice of `bin` are to an
extent a stand-in for `sh` when searched for in an environment that
doesn't have `(git root)/bin` (like the Git for Windows SDK), and
the commands that cannot bve found are to an extent a standi-in for
`sh` on systems where it cannot be found (such as due to `git` not
being installed or installed from MinGit or some other minimal or
nonstandard Git installation, or `GIT_EXEC_PATH` being defined and
having a value that cannot be used or that can be used but points
to a directory structure that does not have a usable `sh`).