Use image:tag naming style for image; let more env through
This changes the style of how we name the custom Docker image for
`cross` testing, so it is not looked up on Docker Hub (it is not
there, and the text before the `/` in the old name was not intended
to have a meaning related to resolving the image in any registry).

This also allows through the environment varibles that currently
have special meaning to gitoxide or its test suite -- including
`GIX_TEST_IGNORE_ARCHIVES` to test the fixture scripts in the
container, if it is set for the `cross` process -- or that the test
suite may reasonably use to detect CI in general or GitHub Actions
in particular. (Most GitHub Actions environment variables continue
not to be passed through, because some of them would hold paths
that couldn't necessarily be resolved properly in the container,
while the presence of others would wrongly create the appearance
that all GitHub Actions environment variables would be usable.)