Don't use leading `./` when there is another `/`
This changes paths that are excuted by a shell in a `justfile` to
omit the leading `./` when the path already has a `/` later.

A path like `./a` is needed to run `a` from the current directory,
but `./a/b` is just a longer way to express `a/b`, since `a/b`
already has a `/` in it. It is the presence of a `/`, rather than
the presence of a leading `./` specifically, that causes paths to
be looked up relative to the current directory rather than by a
`PATH` search.

The reason to make this change is that a few of the commands that
use those paths are about to get a bit more complicated. It is
hoped that removing this small amount of noise, though currently
inconsequential, will allow the immediately forthcoming change to
be made while preserving readability.