Use runner-provided Git for Windows now that it's 2.49.0
Because #1849 (https://github.com/git-for-windows/git/issues/5436)
caused some incorrect test failures in `test-fixtures-windows`, we
have been upgrading Git for Windows on the runner for that job:

- We starting doing that in of 4237e5a (#1870). At that time, we
  upgraded it to the most recent tag, to select the latest release
  even though it was a release candiate, because Git for Windows
  2.49.0 had not yet had a stable release.

- Then Git for Windows 2.49.0 got a stable release, and starting in
  ddef6d3 (#1892) we have upgraded to the release marked "latest"
  (which in Git for Windows is always the latest stable release).

More recently, the Windows Server 2022 runner image (which the
`windows-latest` label we use currently selects) has upgraded its
Git for Windows installation from 2.48.1 to 2.49.0:
https://github.com/actions/runner-images/releases/tag/win22%2F20250330.1

It is therefore no longer necessary for the `test-fixtures-windows`
job to upgrade Git for Windows on the runner. This removes the step
that did that. (Either version of the step could be brought back
from the history if ever needed again.)

This closes #1849, per the plan in:
https://github.com/GitoxideLabs/gitoxide/issues/1849#issuecomment-2724513361