Use the Apache-2.0 license file without the appendix
This removes the appendix from the Apache-2.0 license file. The
license terms themselves are of course not modified. The license
terms end in "END OF TERMS AND CONDITIONS", which is retained, and
the appendix that follows it is removed.

This follows the practice in most or all of the repositories owned
by the Rust organization of using the Apache-2.0 license without
the appendix, on the grounds that it is confusing to keep it, as
the procedure described in the appendix for indicating that source
code is licensed Apache-2.0 is not actually followed in those
repositories: individual source code files are not being given
Apache-2.0 license headers. Note that this is merely an issue of
confusion; the appendix is not part of the terms of the license,
and no permission or limitation depends on whether it is included.

See https://github.com/rust-lang/rust/pull/67734 for details.

Note that the license text itself must of course not be changed.
But this does not change it: the appendix, while it was part of the
license file, is expressly not part of the terms of the license.