Replace BTreeMap by SmallVec
A `SmallVec` provides a better trade-off with respect to performance
than a `BTreeMap` because `suspects` in most cases only contains 1 item.
The cost of calling `BTreeMap::insert` and `BTreeMap::remove` becomes
noticeable in profiling the longer a blame’s history gets and the more
hunks a file is split into.