Remove old unused cargo-deny `license.clarify` entry
3cfd7fa (#1927) removed `LicenseRef-ring` from the `cargo deny`
license allowlist, because we no longer used any `ring` version old
enough to involve the old custom license. But the associated entry
in the `license.clarify` array that definded `LicenseRef-ring` was
not removed, even though it's not needed either given that the
license it clarifies is no longer referenced. This cleans that up.