feat!: `gix-features` uses and republishes `bytesize` version 2
`gix-features` republishes `bytesize`, which has been bumped from
major version 1 to major version 2. Because the interface of
`bytesize` is effectively part of that of `gix-features` due to
the explicit republication of the `bytesize` module in full (with
no documented extra limitations related to interface stability),
this is effectively a breaking change in `gix-features` as well,
though many callers may not be substantially affected.

Major changes that may affect some callers that use `bytesize`
through `gix-features` include the removal of the top-level
`bytesize::to_string()` function, the change in default behavior
from displaying decimal SI units to displaying binary IEC units
(though all or most gitoxide-related creates, in adapting to the
changes, have thus far opted to continue using decimal SI units),
and the small but UI-facing change that decimal SI kilobytes
(units of 1000 bytes) are given with the symbol "kB" rather than
the more ambiguous (and arguably less accurate) symbol "KB".

In addition to republishing `bytesize`, `gix-features` also
republishes `prodash`. Futhermore, some uses of `bytesize` are
transitively through `prodash`, which itself has recently received
an update to use `bytesize` major version 2. (Since `prodash` does
not republish `bytesize`, that is not considered to be a breaking
change in `prodash` itself.) To get the benefits of the newer
version of `bytesize` while avoiding new inconsistencies, and also
to avoid depending on multiple versions of `bytesize`, the
`prodash` dependency version has also been upgraded.

For more information, see:

- https://github.com/GitoxideLabs/gitoxide/discussions/1952
- https://github.com/bytesize-rs/bytesize/releases/tag/bytesize-v2.0.0
- https://github.com/GitoxideLabs/gitoxide/pull/1949
- https://github.com/Byron/prodash/pull/33
- https://github.com/GitoxideLabs/gitoxide/pull/1953