Adapt `Cargo.toml` files in workspace to `gix-features` bump
This does not increment their package versions, which can be done
when they are actually being released. It only bumps the versions
of `gix-features` they specify as dependencies, to allow them to
build again after the "out of band" version bump done in the
preceding commit.

For context, see the preceding commit as well as:
https://github.com/GitoxideLabs/gitoxide/discussions/1952