Test on the new Windows ARM CI runner
In addition to the existing tests on the x86-64 `windows-latest`
runners (which are currently Windows Server 2022), this now also
tests on the new `windows-11-arm` runner, which runs Windows 11 for
ARM64/AArch64.

The main difference that motivates adding these new CI test jobs is
between x86-64 Windows and ARM64 Windows, rather than between
Windows Server 2022 and Windows 11, though the latter difference
might potentially cause the tests to reveal something.

The `windows-11-arm` runners are in preview, and it is possible
that they will not be stable enough to justify keeping these new
jobs as part of matrix definitions with other jobs: one or both of
the new jobs might need to be moved out of its matrix so it can be
made non-blocking for PRs, or even removed altogether.

For details on this preview runner, see:
https://github.blog/changelog/2025-04-14-windows-arm64-hosted-runners-now-available-in-public-preview/