Correctly process overlapping unblamed hunks
Previously, `process_changes` assumed that `UnblamedHunk`s would never
overlap. This constraint has been lifted.

As a consequence, the high-level assumption that two different lines
from a *Blamed File* can never map to the same line in a *Source File*
has been lifted as well. There is not really a way to enforce this
constraint in the blame algorithm alone as it heavily depends on the
algorithm used for diffing.